working/pyproject.toml
[tool.poetry]
name = "working"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
torch = { url = "https://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.0%2Bcu118-cp38-cp38-linux_x86_64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu118/torchaudio-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl" }
transformers = "^4.31.0"
diffusers = "^0.19.3"
accelerate = "^0.21.0"
opencv-python = "^4.8.0.74"
omegaconf = "^2.3.0"
safetensors = "^0.3.1"
controlnet-aux = "^0.0.6"
mediapipe = "^0.10.3"
pillow = "^10.0.0"
super-image = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
